+++ gitmap/constants/constants.go
@@ -12,7 +12,7 @@
 package constants
 
 // Version.
-const Version = "2.28.0"
+const Version = "2.14.0"
 
 // RepoPath is set at build time via -ldflags.
 var RepoPath = ""